implement residual connection
Yes, with just one hidden layer, but it helps. At least when the initial
contribution of the hidden layer is muted down to 0.1. Probably because
this delays the use of the hidden layer.

I've done an ablation check by replacing "*= 0.1" with "*= 0", which
removes the hidden layer and does surprisingly well. (Maybe because of
the recurrent connections outside of the NN.) After ~3000 evaluations it
becomes visible that the ablation is doing worse than the old code.
After ~5000 evaluations the ranking is clear: (new, old, ablation).