fix uninitialized variable
Recently introduced with residual attempt. Really silly. It must have
worked in some binaries, but not in others. Noticed the problem because
the first generation's performance was different on some runs.

Even with all warnings on, gcc won't catch this one. Clang does. Found
by Eigen's range-check (at runtime) if you build in debug mode and are
lucky enough that the value is negative.